A general comment: the responsiveness of the photo carousel swipe and of the + and – buttons at the bottom of the screen is not as easy as I thought it would be…On the test device we have here, the +- buttons are way too close to the bottom of the screen and don’t easily respond to my push.
Tried re-creating but haven’t been able to. At one point, the “W” in
preview was half cut off. 

Facebook says its still not set up yet so can’t test that connection or
see styling of logout button for it, but assume its the same as IG, just
in blue. 

When logging into Instagram, I once got an error message that needs to
be fixed to be spelling “Instagram”, not “InstaGram”. 

On the “Get Your Photos From” page, PHOTOS in the header isn’t centered. 

On the “EDIT” page, the different styles at the bottom of the page
should all be level with the top of the photos creating a straight line
across. 

On the “EDIT” page, the zoom in/zoom out buttons at the top are too
close to “hold to drag & position”

Cart clean dollar amount ($30 vs. $30.45) at top right of screen works
sometimes but other times it has decimal point and 00 after it. Before
checkout, it should never have a decimal – all our products are clean
dollar amounts. 

At checkout, when choosing “Ship Within USA” option, we should say
something about how long it takes to produce/receive. At the bottom,
right between “Total With Shipping: $xx.xx” and CHECKOUT button, we
should add “You will receive your order in 5-7 business days.” 

We need to finalize the app name and icon. What do we have them as right
now?